requirements.txt
nicegui
Pillow
reportlab